+++ changelog.md
@@ -3,6 +3,7 @@
 ## What's Changed
 
 - Add ENERGY_COUNTER_FEED_IN
+- Fix backend detection
 - Fix install mode for HmIP
 
 # Version 2025.12.5 (2025-12-03)